loading time optimization
probably removed all loading time

+++ module/common/get_extra.sh
@@ -7,35 +7,33 @@ KBOUTPUT="$MODPATH/.extra"
 
 . $MODPATH/util_func.sh
 
-find_busybox
 check_wget
 
-# Fetch additional package names
-wget --no-check-certificate -q -O - "https://raw.githubusercontent.com/KOWX712/Tricky-Addon-Update-Target-List/main/more-excldue.json" 2>/dev/null | \
-grep -o '"package-name": *"[^"]*"' | \
-awk -F'"' '{print $4}' > "$OUTPUT"
-
-if [ ! -s "$OUTPUT" ]; then
-    rm -f "$KBOUTPUT"
-    skipfetch=true
-fi
-
-# Find xposed package name
-pm list packages -3  </dev/null 2>&1 | cat | awk -F: '{print $2}' | while read -r PACKAGE; do
-    if ! grep -Fq "$PACKAGE" "$SKIPLIST"; then
-        pm path "$PACKAGE" | grep "base.apk" | awk -F: '{print $2}' | tr -d '\r' | while read -r APK_PATH; do
-            aapt dump xmltree "$APK_PATH" AndroidManifest.xml 2>/dev/null | grep -qE "xposed.category|xposeddescription" && echo "$PACKAGE" >> "$OUTPUT"
-        done
-    fi
-done
-
-if [ "$skipfetch" != "true" ]; then
+get_kb() {
     wget --no-check-certificate -qO "$KBOUTPUT" "https://raw.githubusercontent.com/KOWX712/Tricky-Addon-Update-Target-List/main/.extra"
+    [ -s "$KBOUTPUT" ] || rm -f "$KBOUTPUT"
+}
 
-    if [ ! -s "$KBOUTPUT" ]; then
-        rm -f "$KBOUTPUT"
+get_unnecessary() {
+    if [ ! -s "$OUTPUT" ] || [ ! -f "$OUTPUT" ]; then
+        wget --no-check-certificate -q -O - "https://raw.githubusercontent.com/KOWX712/Tricky-Addon-Update-Target-List/main/more-excldue.json" 2>/dev/null | \
+        grep -o '"package-name": *"[^"]*"' | \
+        awk -F'"' '{print $4}' > "$OUTPUT"
     fi
+    get_xposed
+}
 
+get_xposed() {
+    pm list packages -3 | awk -F: '{print $2}' | while read -r PACKAGE; do
+        if ! grep -Fq "$PACKAGE" "$SKIPLIST" && ! grep -Fq "$PACKAGE" "$OUTPUT"; then
+            pm path "$PACKAGE" | grep "base.apk" | awk -F: '{print $2}' | tr -d '\r' | while read -r APK_PATH; do
+                aapt dump xmltree "$APK_PATH" AndroidManifest.xml 2>/dev/null | grep -qE "xposed.category|xposeddescription" && echo "$PACKAGE" >> "$OUTPUT"
+            done
+        fi
+    done
+}
+
+check_update() {
     if [ -d "$MODPATH/temp" ]; then
         JSON=$(wget --no-check-certificate -q -O - "https://raw.githubusercontent.com/KOWX712/Tricky-Addon-Update-Target-List/main/update.json")
         REMOTE_VERSION=$(echo "$JSON" | grep -o '"versionCode": *[0-9]*' | awk -F: '{print $2}' | tr -d ' ')
@@ -44,6 +42,11 @@ if [ "$skipfetch" != "true" ]; then
             echo "update"
         fi
     fi
-else
-    exit 1
-fi
+}
+
+case "$1" in 
+	--kb) get_kb; exit ;;
+	--unnecessary) get_unnecessary; exit ;;
+	--xposed) get_xposed; exit ;;
+	--update) check_update; exit ;;
+esac
